Folds skills back into About Body
